# backend/internal/httpapi/openapi.yaml
openapi: 3.1.0
info:
  title: Home Finance API
  version: "1.4.0"
  description: |
    Internal API for a personal/family budgeting app. Plain net/http + chi
    router, hand-written JSON handlers. All money amounts are integer cents.
    Dates are YYYY-MM-DD strings; timestamps are RFC3339.

    Every error response uses the shape
    `{"error": {"code": "...", "message": "..."}}`. Messages are safe to show
    to a user; `db_error` carries a fixed generic message and never raw driver
    text.

    JSON request bodies are capped at 1 MiB by middleware; a larger body gets
    413 `body_too_large`. Multipart uploads bypass that cap and enforce their
    own, larger limit.

    Unmatched `/api` paths return 404 `not_found` and unsupported methods
    return 405 `method_not_allowed`, both as JSON rather than the SPA.

    Operations marked "Admin only" also require the caller's role to be
    `admin`; other roles get 403 `forbidden`.
servers:
  - url: /
...
security:
  - cookieAuth: []
    csrfHeader: []
paths:
...
  /api/kids/balances:
    get:
      operationId: getKidBalances
      tags: [Kids]
      summary: Lists non-archived kids with their ours/theirs/total ledger balances and last reported balance.
      responses:
        "200":
          description: Kid balances.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KidBalance"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/kids:
    get:
      operationId: listKids
      tags: [Kids]
      summary: Lists non-archived kids, ordered by sort order.
      responses:
        "200":
          description: Kids.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Kid"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/kids/{id}/reported-balance:
    patch:
      operationId: updateKidReportedBalance
      tags: [Kids]
      summary: Records the balance the bank reports for a kid, for reconciliation. Admin only. Sending a null reportedBalanceCents clears both fields.
      parameters:
        - $ref: "#/components/parameters/EntityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKidReportedBalanceRequest"
      responses:
        "204":
          description: Reported balance updated.
        "400":
          description: Invalid id, amount out of range, or reportedBalanceDate missing while reportedBalanceCents is set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Kid not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/kids/{id}/ledger:
    get:
      operationId: getKidLedger
      tags: [Kids]
      summary: Lists a kid's savings ledger in date order, with running ours/theirs balances.
      parameters:
        - $ref: "#/components/parameters/EntityId"
      responses:
        "200":
          description: Kid ledger entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KidLedgerEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/kids/{id}/entries:
    post:
      operationId: createKidLedgerEntry
      tags: [Kids]
      summary: Adds a ledger entry to a kid. Withdrawal amounts are coerced negative. The response carries the resolved entryDate - today's date when the request omitted it.
      parameters:
        - $ref: "#/components/parameters/EntityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKidLedgerEntryRequest"
      responses:
        "201":
          description: Entry created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KidEntryCreated"
        "400":
          description: Invalid id, owner, entryType, amount out of range, or entryDate not YYYY-MM-DD.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Kid not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: "`year_locked`: the entry's calendar year is locked."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/kid-entries/{id}:
    patch:
      operationId: updateKidLedgerEntry
      tags: [Kids]
      summary: Updates a kid ledger entry's owner and amount. Withdrawal amounts are coerced negative.
      parameters:
        - $ref: "#/components/parameters/EntityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKidLedgerEntryRequest"
      responses:
        "200":
          description: Entry updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KidEntryUpdated"
        "400":
          description: Invalid id, owner, or amount out of range.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Entry not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: "`year_locked`: the entry's calendar year is locked."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      operationId: deleteKidLedgerEntry
      tags: [Kids]
      summary: Deletes a kid ledger entry.
      parameters:
        - $ref: "#/components/parameters/EntityId"
      responses:
        "204":
          description: Entry deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Entry not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: "`year_locked`: the entry's calendar year is locked."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie set after successful login (OIDC or dev fake auth).
    csrfHeader:
      type: apiKey
      in: header
      name: X-Requested-With
      description: Required (value "XMLHttpRequest") on all mutating /api requests alongside the session cookie.
  parameters:
...
    EntityId:
      name: id
      in: path
      required: true
      description: Resource id.
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Invalid request (code `bad_request`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Admin role required, or the X-Requested-With CSRF header is missing (code `forbidden` / `csrf`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BodyTooLarge:
      description: JSON body exceeded the 1 MiB limit (code `body_too_large`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ServerError:
      description: Server or database error (code `db_error` / `server_error`); the message is generic.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
...
    Kid:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        sortOrder:
          type: integer
        archivedAt:
          type: string
          format: date-time
        reportedBalanceCents:
          type: integer
          format: int64
        reportedBalanceDate:
          type: string
          format: date
      required: [id, name, sortOrder]
    KidBalance:
      type: object
      description: Row from GET /api/kids/balances - non-archived kids only.
      properties:
        kidId:
          type: integer
          format: int64
        name:
          type: string
        oursCents:
          type: integer
          format: int64
        theirsCents:
          type: integer
          format: int64
        totalCents:
          type: integer
          format: int64
        reportedBalanceCents:
          type: integer
          format: int64
        reportedBalanceDate:
          type: string
          format: date
      required: [kidId, name, oursCents, theirsCents, totalCents]
    KidLedgerEntry:
      type: object
      description: Row from GET /api/kids/{id}/ledger.
      properties:
        id:
          type: integer
          format: int64
        kidId:
          type: integer
          format: int64
        owner:
          type: string
          enum: [ours, theirs]
        entryType:
          type: string
          enum: [opening_balance, deposit, withdrawal, adjustment]
        amountCents:
          type: integer
          format: int64
        runningOursCents:
          type: integer
          format: int64
        runningTheirsCents:
          type: integer
          format: int64
        description:
          type: string
        entryDate:
          type: string
          format: date
      required:
        - id
        - kidId
        - owner
        - entryType
        - amountCents
        - runningOursCents
        - runningTheirsCents
        - description
        - entryDate
    CreateKidLedgerEntryRequest:
      type: object
      properties:
        owner:
          type: string
          enum: [ours, theirs]
        entryType:
          type: string
          enum: [opening_balance, deposit, withdrawal, adjustment]
        amountCents:
          type: integer
          format: int64
          description: For entryType=withdrawal, a positive value is coerced to negative server-side.
        description:
          type: string
        entryDate:
          type: [string, "null"]
          format: date
          description: Defaults to today when omitted; the locked-year check is applied to this date.
      required: [owner, entryType, amountCents]
    KidEntryCreated:
      type: object
      description: Response from POST /api/kids/{id}/entries.
      properties:
        id:
          type: integer
          format: int64
        kidId:
          type: integer
          format: int64
        owner:
          type: string
          enum: [ours, theirs]
        entryType:
          type: string
          enum: [opening_balance, deposit, withdrawal, adjustment]
        amountCents:
          type: integer
          format: int64
        description:
          type: string
        entryDate:
          type: string
          format: date
          description: The date actually stored - the submitted value, or today when it was omitted.
      required: [id, kidId, owner, entryType, amountCents, description, entryDate]
    UpdateKidLedgerEntryRequest:
      type: object
      properties:
        owner:
          type: string
          enum: [ours, theirs]
        amountCents:
          type: integer
          format: int64
      required: [owner, amountCents]
    KidEntryUpdated:
      type: object
      description: Response from PATCH /api/kid-entries/{id}.
      properties:
        id:
          type: integer
          format: int64
        kidId:
          type: integer
          format: int64
        owner:
          type: string
          enum: [ours, theirs]
        entryType:
          type: string
          enum: [opening_balance, deposit, withdrawal, adjustment]
        amountCents:
          type: integer
          format: int64
      required: [id, kidId, owner, entryType, amountCents]
    UpdateKidReportedBalanceRequest:
      type: object
      properties:
        reportedBalanceCents:
          type: [integer, "null"]
          format: int64
          description: Null clears both fields.
        reportedBalanceDate:
          type: [string, "null"]
          format: date
          description: Required whenever reportedBalanceCents is set.
...